refactor(earn/neeru): replace broken manage bottom sheet with a full screen
The @gorhom/bottom-sheet-based NeeruCloseSheet + NeeruEmergencyCloseSheet
never reliably rose from either the MyPools card path or the AllPools
detail path. Symptoms: user tapped Administrar and the sheet either did
nothing or "tried to rise but stayed at 0 height". Multiple attempts to
fix the layout race (prime-on-mount + snapToIndex, explicit snapPoints,
BottomSheetView vs BottomSheetScrollView, hosting at screen level) did
not help in the virtualized-item / persisted-state combination we have
here.

Ship a plain full-screen NeeruManagePosition instead. Every navigation
in the app already works, so this has zero layout-timing risk.

The screen carries the same lifecycle guarantees the sheet had:
- Confirmar retiro dispatches closePositionStart, saga handles the tx.
- If the fund can't pay interest (NEERU_LOW_POOL_ERROR) the emergency
  fallback surfaces inline: warning copy + "Retirar solo el deposito"
  button that dispatches emergencyCloseStart. Mount-scoped so stale
  Redux state from a previous session cannot show it on open.
- On success, navigateBack() pops back to the caller (mount-scoped
  guard prevents stale closeStatus='success' from auto-dismissing on
  arrival).
- Depositar (renombrado "Haz otro deposito" per user ask, cada
  posicion es independiente) navigates to EarnEnterAmount.

Both entry points now route the same way:
- MyPools > NeeruPositionCard "Administrar" -> NeeruManagePosition
- AllPools > NeeruVaultDetail > NeeruPositionRow "Administrar" ->
  NeeruManagePosition

Deleted (no longer referenced): NeeruCloseSheet, NeeruEmergencyCloseSheet
plus their test files. NeeruVaultDetailScreen test updated to assert
navigation instead of sheet mount.
